Make EpochId 64-bit and SlotId 16-bit
This makes them correspond to the types used in the Haskell
implementation.

+++ storage/src/block/iter.rs
@@ -108,7 +108,7 @@ pub fn resolve_date_to_blockhash(storage: &Storage, tip: &BlockHash, date: &Bloc
                 BlockDate::Genesis(_) => 0,
                 BlockDate::Normal(sid) => sid.slotid,
             };
-            let r = handle.getref_at_index(slotid)?;
+            let r = handle.getref_at_index(slotid as u32)?;
             Ok(r)
         },
         Err(_) => {
